git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@5019 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/FILTERING/TRANSFORMERS/LinearResampler.h
@@ -85,7 +85,7 @@ namespace OpenMS
       	
         double end_pos = (last-1)->getMZ();
         double start_pos = first->getMZ();
-        int number_raw_points = spectrum.size();
+        int number_raw_points = (int)spectrum.size();
         int number_resampled_points = (int)(ceil((end_pos -start_pos) / spacing_ + 1));
 
         typename std::vector<PeakType> resampled_peak_container;
